moved new widget to `histogram.py`
